feat(cli): Add --log flag for verbose output
This change introduces a new global  flag that allows users to
control the log level of the application. The flag accepts one of the
following values: error, warn, info, debug, trace. The default log
level is info.

This feature addresses the customer request for more detailed output
for debugging and monitoring purposes.